+++ firmware/examples/zedboard/src/bin/l3gd20h-spi-mio.rs
@@ -18,18 +18,16 @@ use log::{error, info};
 use zynq7000_hal::{
     BootMode,
     clocks::Clocks,
-    configure_level_shifter, generic_interrupt_handler,
-    gic::Configurator,
+    generic_interrupt_handler,
     gpio::{GpioPins, Output, PinState},
     gtc::GlobalTimerCounter,
-    l2_cache,
     log::asynch::UartLoggerRunner,
     spi::{self, SpiAsync, SpiWithHwCs, SpiWithHwCsAsync},
     time::Hertz,
     uart::{self, TxAsync},
 };
 
-use zynq7000::{Peripherals, slcr::LevelShifterConfig, spi::DelayControl};
+use zynq7000::spi::DelayControl;
 use zynq7000_rt as _;
 
 // Define the clock frequency as a constant
@@ -46,11 +44,7 @@ fn entry_point() -> ! {
 
 #[embassy_executor::main]
 async fn main(spawner: Spawner) -> ! {
-    let mut dp = Peripherals::take().unwrap();
-    l2_cache::init_with_defaults(&mut dp.l2c);
-
-    // Enable PS-PL level shifters.
-    configure_level_shifter(LevelShifterConfig::EnableAll);
+    let periphs = zynq7000_hal::init(zynq7000_hal::Config::default()).unwrap();
 
     // Clock was already initialized by PS7 Init TCL script or FSBL, we just read it.
     let mut clocks = Clocks::new_from_regs(PS_CLOCK_FREQUENCY).unwrap();
@@ -67,19 +61,10 @@ async fn main(spawner: Spawner) -> ! {
         "SPI reference clock must be larger than CPU 1x clock"
     );
 
-    // Set up the global interrupt controller.
-    let mut gic = Configurator::new_with_init(dp.gicc, dp.gicd);
-    gic.enable_all_interrupts();
-    gic.set_all_spi_interrupt_targets_cpu0();
-    gic.enable();
-    unsafe {
-        gic.enable_interrupts();
-    }
-
-    let mut gpio_pins = GpioPins::new(dp.gpio);
+    let mut gpio_pins = GpioPins::new(periphs.gpio);
 
     // Set up global timer counter and embassy time driver.
-    let gtc = GlobalTimerCounter::new(dp.gtc, clocks.arm_clocks());
+    let gtc = GlobalTimerCounter::new(periphs.gtc, clocks.arm_clocks());
     zynq7000_hal::time_driver_gtc::init(clocks.arm_clocks(), gtc);
 
     // Set up the UART, we are logging with it.
@@ -87,7 +72,7 @@ async fn main(spawner: Spawner) -> ! {
         .unwrap()
         .0;
     let mut uart = uart::Uart::new_with_mio_for_uart_1(
-        dp.uart_1,
+        periphs.uart_1,
         uart::Config::new_with_clk_config(uart_clk_config),
         (gpio_pins.mio.mio48, gpio_pins.mio.mio49),
     )
@@ -112,7 +97,7 @@ async fn main(spawner: Spawner) -> ! {
     }
 
     let mut spi = spi::Spi::new_one_hw_cs(
-        dp.spi_1,
+        periphs.spi_1,
         spi::Config::new(
             // 10 MHz maximum rating of the sensor.
             zynq7000::spi::BaudDivSel::By64,