+++ firmware/examples/zedboard/src/bin/uart-blocking.rs
@@ -13,14 +13,12 @@ use zedboard::PS_CLOCK_FREQUENCY;
 use zynq7000_hal::{
     BootMode,
     clocks::Clocks,
-    configure_level_shifter, generic_interrupt_handler, gic,
+    generic_interrupt_handler,
     gpio::{GpioPins, Output, PinState},
     gtc::GlobalTimerCounter,
-    l2_cache,
     uart::{ClockConfig, Config, Uart},
 };
 
-use zynq7000::{Peripherals, slcr::LevelShifterConfig};
 use zynq7000_rt as _;
 
 const INIT_STRING: &str = "-- Zynq 7000 Zedboard blocking UART example --\n\r";
@@ -94,25 +92,14 @@ impl UartMultiplexer {
 }
 #[embassy_executor::main]
 async fn main(_spawner: Spawner) -> ! {
-    let mut dp = Peripherals::take().unwrap();
-    l2_cache::init_with_defaults(&mut dp.l2c);
+    let periphs = zynq7000_hal::init(zynq7000_hal::Config::default()).unwrap();
 
-    // Enable PS-PL level shifters.
-    configure_level_shifter(LevelShifterConfig::EnableAll);
     // Clock was already initialized by PS7 Init TCL script or FSBL, we just read it.
     let clocks = Clocks::new_from_regs(PS_CLOCK_FREQUENCY).unwrap();
-    // Set up the global interrupt controller.
-    let mut gic = gic::Configurator::new_with_init(dp.gicc, dp.gicd);
-    gic.enable_all_interrupts();
-    gic.set_all_spi_interrupt_targets_cpu0();
-    gic.enable();
-    unsafe {
-        gic.enable_interrupts();
-    }
-    let mut gpio_pins = GpioPins::new(dp.gpio);
+    let mut gpio_pins = GpioPins::new(periphs.gpio);
 
     // Set up global timer counter and embassy time driver.
-    let gtc = GlobalTimerCounter::new(dp.gtc, clocks.arm_clocks());
+    let gtc = GlobalTimerCounter::new(periphs.gtc, clocks.arm_clocks());
     zynq7000_hal::time_driver_gtc::init(clocks.arm_clocks(), gtc);
 
     // Set up the UART, we are logging with it.
@@ -120,7 +107,7 @@ async fn main(_spawner: Spawner) -> ! {
         .unwrap()
         .0;
     let mut log_uart = Uart::new_with_mio_for_uart_1(
-        dp.uart_1,
+        periphs.uart_1,
         Config::new_with_clk_config(uart_clk_config),
         (gpio_pins.mio.mio48, gpio_pins.mio.mio49),
     )
@@ -132,7 +119,7 @@ async fn main(_spawner: Spawner) -> ! {
 
     // UART0 routed through EMIO to PL pins.
     let mut uart_0 =
-        Uart::new_with_emio(dp.uart_0, Config::new_with_clk_config(uart_clk_config)).unwrap();
+        Uart::new_with_emio(periphs.uart_0, Config::new_with_clk_config(uart_clk_config)).unwrap();
     // Safety: Valid address of AXI UARTLITE.
     let mut uartlite = unsafe { AxiUartlite::new(AXI_UARTLITE_BASE_ADDR) };